+++ workers/loc.api/helpers/get-rest.js
@@ -1,0 +1,16 @@
+'use strict'
+
+const bfxFactory = require('./bfx.factory')
+
+module.exports = (auth, wrkReportServiceApi) => {
+  if (typeof auth !== 'object') {
+    throw new Error('ERR_AUTH_UNAUTHORIZED')
+  }
+
+  const group = wrkReportServiceApi.group
+  const conf = wrkReportServiceApi.conf[group]
+
+  const bfx = bfxFactory({ conf, ...auth })
+
+  return bfx.rest(2, { transform: true })
+}
