+++ workers/loc.api/helpers/has-job-in-queue-with-status-by.js
@@ -1,0 +1,53 @@
+'use strict'
+
+const { promisify } = require('util')
+
+module.exports = async (
+  reportService,
+  args,
+  statuses = ['ACTIVE', 'PROCESSING']
+) => {
+  const ctx = reportService.ctx
+  const wrk = ctx.grc_bfx.caller
+  const group = wrk.group
+  const conf = wrk.conf[group]
+
+  if (
+    conf.syncMode ||
+    !conf.isSpamRestrictionMode
+  ) {
+    await promisify(reportService.getEmail.bind(reportService))(null, args)
+
+    return null
+  }
+
+  const userInfo = await reportService._getUserInfo(args)
+
+  const procQ = ctx.lokue_processor.q
+  const aggrQ = ctx.lokue_aggregator.q
+
+  const hasJobInQueue = !(statuses.every(status => {
+    return [procQ, aggrQ].every(queue => {
+      const jobs = queue.listJobs(status)
+
+      if (!Array.isArray(jobs)) {
+        return true
+      }
+
+      return jobs.every(job => {
+        return (
+          typeof job === 'object' &&
+          typeof job.data === 'object' &&
+          typeof job.data.userId !== 'undefined' &&
+          job.data.userId !== userInfo.id
+        )
+      })
+    })
+  }))
+
+  if (hasJobInQueue) {
+    throw new Error('ERR_HAS_JOB_IN_QUEUE')
+  }
+
+  return userInfo.id
+}
